test: mark `configure_output` tests as `#[serial]`
`configure_output` manipulates `yansi::Color`'s global enable toggle;
this state is [not thread local](https://docs.rs/yansi/0.5.1/src/yansi/paint.rs.html#373-374)
so changes to it affect the whole program.

This causes the `configure_output` tests to fail (sporadically) when
Cargo chooses to run tests in parallel and when the interleaving of
`configure_output` tests results in `Paint::enable` being called between
`configure_output` and `.paint` in another test.

Because we're dealing with a race condition there isn't a great way to
reproduce the failures but running:
```console
> until ! cargo test --jobs 10 ; do true; done
```

Reliably fails for me within a few runs.

---

Marking these tests with `#[serial]` (from the
[`serial_test` crate](https://docs.rs/serial_test/2.0.0/serial_test/index.html))
fixes this.